requirements.txt
# Core dependencies
requests>=2.31.0
slack-sdk>=3.21.0
openai>=1.3.0
feedparser>=6.0.10
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-dateutil>=2.8.0
pytz>=2023.3

# Database
# sqlite3 is part of Python's standard library - no need to install

# Configuration
PyYAML>=6.0

# Text processing
nltk>=3.8.1
textblob>=0.17.1
langdetect>=1.0.9

# Web scraping
selenium>=4.15.0
webdriver-manager>=4.0.0

# Social media APIs
tweepy>=4.14.0

# Scheduling
APScheduler>=3.10.0

# Logging and monitoring
structlog>=23.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0

# Development
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0

# Environment management
python-dotenv>=1.0.0

# HTTP client with async support
aiohttp>=3.8.0
httpx>=0.24.0

# Data validation
pydantic>=2.0.0

# Utilities
click>=8.1.0
rich>=13.5.0
